# 🐰 Wrap code in functions to make it faster