--- a/sample/Interactivity.jl
+++ b/sample/Interactivity.jl
@@ -165,6 +165,26 @@ In fact, **_any package_ can add bindable values to their objects**. For example
 
 A package _does not need to add `Pluto.jl` as a dependency to do so_: only the `Base.show(io, MIME("text/html"), obj)` function needs to be extended to contain a `<script>` that triggers the `input` event with a value. (It's up to the package creator _when_ and _what_.) This _does not affect_ how the object is displayed outside of Pluto.jl: uncaught events are ignored by your browser."""
 
+# ╔═╡ aa8f6a0e-303a-11eb-02b7-5597c167596d
+
+
+# ╔═╡ 5c1ececa-303a-11eb-1faf-0f3a6f94ac48
+md"""## Separate definition and reference
+Interactivity works through reactivity. If you put a bond and a reference to the same variable together, then setting the bond will trigger the _entire cell_ to re-evaluate, including the bond itself.
+
+So **do not** write
+```julia
+md""\"$(@bind r html"<input type=range>")  $(r^2)""\"
+```
+Instead, create two cells:
+```julia
+md""\"$(@bind r html"<input type=range>")""\"
+```
+```julia
+r^2
+```
+"""
+
 # ╔═╡ 55783466-7eb1-11ea-32d8-a97311229e93
 
 
@@ -245,6 +265,8 @@ md"That's it for now! Let us know what you think using the feedback button below
 # ╟─d774fafa-7f34-11ea-290d-37805806e14b
 # ╟─8db857f8-7eae-11ea-3e53-058a953f2232
 # ╟─d5b3be4a-7f52-11ea-2fc7-a5835808207d
+# ╟─aa8f6a0e-303a-11eb-02b7-5597c167596d
+# ╟─5c1ececa-303a-11eb-1faf-0f3a6f94ac48
 # ╟─55783466-7eb1-11ea-32d8-a97311229e93
 # ╟─582769e6-7eb1-11ea-077d-d9b4a3226aac
 # ╟─8f829274-7eb1-11ea-3888-13c00b3ba70f