--- a/src/analysis/ExpressionExplorer.jl
+++ b/src/analysis/ExpressionExplorer.jl
@@ -928,10 +928,13 @@ function can_be_function_wrapped(x::Expr)
         x.head === :function ||
         x.head === :macro ||
         x.head === :macrocall ||
-        x.head === :macro
+        x.head === :macro ||
+        x.head === :struct ||
+        x.head === :abstract ||
+        (x.head === :(=) && x.args[1] isa Expr && x.args[1].head === :call)
         false
     else
-        true
+        all(can_be_function_wrapped, x.args)
     end
 
 end