--- a/test/React.jl
+++ b/test/React.jl
@@ -816,8 +816,6 @@ import Distributed
         @test occursinerror("You can only use return inside a function.", notebook.cells[3])
         @test occursinerror("You can only use return inside a function.", notebook.cells[4])
         @test occursinerror("You can only use return inside a function.", notebook.cells[5])
-        @test notebook.cells[1].output_repr == 1
-        @test notebook.cells[5].output_repr == 1
 
         @test notebook.cells[6].errored == false