We turn a simple cell like `p = plot(x,y)` into a _function_ so that it can be called a second time without compilation!

It takes the referenced globals as arguments, and returns the definined globals. `p = plot(x,y)` would become:

```julia
function(plot, x,y)
	result = p = plot(x,y)
	(result, (p,))
end
```

Function generation is here:
https://github.com/fonsp/Pluto.jl/pull/720/files#diff-0cc97f3d6a0f647a05e5f913d416242ecb00f6e67e12004a7204b8846a3fa44cR96-R104

and it is called here:
https://github.com/fonsp/Pluto.jl/pull/720/files#diff-0cc97f3d6a0f647a05e5f913d416242ecb00f6e67e12004a7204b8846a3fa44cR112

---

Not every cell will be run as a function. Cells that import, define a function, or type, or call a macro still run with the old method:
https://github.com/fonsp/Pluto.jl/pull/720/files#diff-04c7252262a0e6de2db2cded7aba9b1130c1e5ba6adff25bf8ec3a34e36a4788R928-R946

---

And the best part is: **all of this is completely invisible**! You get superfast code without having to rewrite anything ❤