@fonsp 2020-11-26
Right now I did the rewrite but no cell is being function wrapped yet