@fonsp 2020-11-26
TODO:
- [ ] Remove the tip to wrap your code inside a function in the Interactivity notebook 😊
- [ ] Measure performance boost
- [ ] Measure increase in RAM usage?