@lungben 2020-11-27
looks good so far!

a)
```
function do_stuff(x)
	y = 0.0
	for i = 1:100000000
		y += x/i
	end
	y
end

do_stuff(x)
```

b)
```
begin
	y = 0.0
	for i = 1:100000000
		global y += x/i # global keyword may be obmitted in this PR because of implicit function wrapping
	end
	y
end
```

Timings:
without this PR:
a) ~112 ms
b) 10s (`global y` is required)

with this PR:
a) ~112 ms
b) 10s (with `global y`) / 112 ms (without `global y`)

Observations:
* I tried a few of my own notebooks and nothing crashed :-)
* if just a function is executed in a cell, the timings are the same (especially, no performance loss)
* the scoping rules change for loops inside a cell because they are now implicitly inside a function - this may be unexpected, but is not bad imho, it is essentially the same as the new Julia 1.5 REPL scoping (which is currently not used in Pluto).
* removing the global keyword in the loop gives with this PR the same performance (a factor of 1000 better!) as manually wrapping it into a function. 

Suggestions:
* It would be good to have a way to define "local" variables, which are not returned to global scope. This could be helpful when dealing with large amount of data to allow GC to clean up intermediate results. We could use the Python convention of variables starting with underscore (e.g. `_x`) for private variables which are not exported. This would somehow complement let blocks, where only the return values are exported and everything else is local.
* A blog post / discourse, etc. on this change would be helpful - understanding this change allows users to write more efficient notebooks (e.g. by using function scoping instead of globals).