@fonsp 2020-11-27
About local variables: this is why Julia has `local`: https://github.com/fonsp/Pluto.jl/issues/379#issuecomment-689130899