@lungben 2020-11-27
Thanks!
Using `local` for `y` the code runs in 8s without this PR and 112 ms with this PR 😄 
Furthermore, the `local` keyword prevents the variable from being exported to global scope, therefore my suggestion above is already possible without me recognizing it.