@fonsp 2020-11-27
That's right, 40x faster than the REPL!

![image](https://user-images.githubusercontent.com/6933510/100488019-03ea7e00-310c-11eb-8efb-2628755cd257.png)