@NHDaly 2020-12-15
Also, wow, it's so cool how _much_ faster this got.

Hehe technically this isn't including the compilation time in the time result, but that's standard for what julia does for `@time` (to time the _actual_ time including compilation, you can time `@eval`, which can be important if your code actually takes potentially several minutes to compile and then time reports it ran almost instantly):
```julia
julia> @time 2+2
  0.000000 seconds
4

julia> @time @eval 2+2
  0.000970 seconds (53 allocations: 3.516 KiB)
4
```

But still, after this change, even if you include compilation time, things get much faster because you've changed a non-const global into a function argument. For the old case before this PR, julia had to check every iteration of the loop that `y`'s value hadn't changed (because it's just a global, and it could have been affected by e.g. another thread somewhere or something), and in the new case, it can rely on the value being a constant, which matches the reactive notebook paradigm much better! :)

So even if you include the compilation time, the new system is still faster than evaling the block old way! 🎉 
<img width="734" alt="Screen Shot 2020-12-15 at 11 52 49 AM" src="https://user-images.githubusercontent.com/1582097/102246659-03543480-3ecd-11eb-83b1-4ff257914105.png">
<img width="728" alt="Screen Shot 2020-12-15 at 11 53 49 AM" src="https://user-images.githubusercontent.com/1582097/102246677-064f2500-3ecd-11eb-9132-e30fabc09460.png">


------

I guess Pluto could consider to separately report the compilation time as well? Julia 1.6 has started doing that: 
https://github.com/JuliaLang/julia/pull/37678
(Although they are missing some "top-level" compilation time, as reported here: https://github.com/JuliaLang/julia/pull/37938. Pluto is better positioned to get this report correct, if that's desirable, but i'm not sure whether it is or isn't. 🙂)